Create creditCardGenerator
This C program serves as a Credit Card Number Generator. It prompts the user to specify the number of credit card numbers they want to generate. Using a random number generation algorithm, it generates the specified number of 16-digit credit card numbers (in a simplified format) and displays them. The program also incorporates a timestamp-based seed for the random number generator to ensure different sequences of numbers on each run.