Create ramanujanNumber
This C program identifies and prints Ramanujan numbers up to a specified limit. A Ramanujan number is a number that can be expressed as the sum of two cubes in two different ways. The program computes the cube values for numbers up to the given limit and then checks each number within that limit to determine if it qualifies as a Ramanujan number based on the sum of two cubes. If a number meets this criteria, it is displayed as a Ramanujan number.